Update magtag_quote_demo.py
testing with magtag 2025 and the colors are coming up inverted if set directly (bg is black and text is white). when i remove them then the colors are correct. the other examples leave text_color and default_bg as default. not sure if this is an issue though with the new display and how it is init in the core or in portalbase/this library because setting the color i imagine should work